Former-commit-id: f08ccbf9b0fa138a37e8aa17627180b3e923643b

+++ modules/SofaBaseCollision/CubeModel.h
@@ -116,7 +116,7 @@ public:
     void setLeafCube(int cubeIndex, std::pair<core::CollisionElementIterator,core::CollisionElementIterator> children, const sofa::defaulttype::Vector3& min, const sofa::defaulttype::Vector3& max);
 
 
-    unsigned int getNumberCells() { return elems.size();}
+    unsigned int getNumberCells() { return (unsigned int)elems.size();}
 
     void getBoundingTree ( sofa::helper::vector< std::pair< sofa::defaulttype::Vector3, sofa::defaulttype::Vector3> > &bounding )
     {
